feat(modelos): Refactor LogisticaInputInserter to handle Demanda and Oferta inputs

See merge request apps/operaciones-industriales/blending-harina-2/blending-harina-2-backend!36

+++ Function.Blending.Opt/Application/Features/LogEjecucion/DTOs/Requests/Input/LogInpFilCapacidadDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace Function.Blending.Opt.Application.Features.LogEjecucion.DTOs.Requests.Input;
 
 public sealed record LogInpFilCapacidadDto(
-  int Cantidad,
-  int Capacidad
+  decimal Cantidad,
+  decimal Capacidad
 );
