+++ Function.Blending.Opt/Domain/ValueObjects/LogOutConDistribucion.cs
@@ -4,5 +4,5 @@ public sealed record LogOutConDistribucion(
   Guid? Id = default,
   Guid? ContenedorId = default,
   string? Ruma = null,
-  int? Valor = null
+  decimal? Valor = null
 );
